+++ AuthenticatorShared/UI/Platform/Application/Extensions/View.swift
@@ -60,24 +60,6 @@ extension View {
         }
     }
 
-    /// Hides a view based on the specified value.
-    ///
-    /// NOTE: This should only be used when the view needs to remain in the view hierarchy while hidden,
-    /// which is often useful for sizing purposes (e.g. hide or swap a view without resizing the parent).
-    /// Otherwise, `if condition { view }` is preferred.
-    ///
-    /// - Parameter hidden: `true` if the view should be hidden.
-    /// - Returns The original view if `hidden` is false, or the view with the hidden modifier applied.
-    ///
-    @ViewBuilder
-    func hidden(_ hidden: Bool) -> some View {
-        if hidden {
-            self.hidden()
-        } else {
-            self
-        }
-    }
-
     /// Applies a custom navigation bar title and title display mode to a view.
     ///
     /// - Parameters:
