+++ BitwardenKit/UI/Platform/Application/Extensions/ModifiedContent+Conditionals.swift
