+++ BitwardenShared/UI/Platform/Application/Extensions/View.swift
@@ -5,13 +5,6 @@ import SwiftUI
 /// Helper functions extended off the `View` protocol.
 ///
 extension View {
-    /// Apply an arbitrary block of modifiers to a view. This is particularly useful
-    /// if the modifiers in question might only be available on particular versions
-    /// of iOS.
-    func apply<V: View>(@ViewBuilder _ block: (Self) -> V) -> V {
-        block(self)
-    }
-
     /// Focuses next field in sequence, from the given `FocusState`.
     /// Requires a currently active focus state and a next field available in the sequence.
     /// (https://stackoverflow.com/a/71531523)
@@ -37,38 +30,6 @@ extension View {
         let nextValue = currentValue.rawValue + 1
         if let newValue = F(rawValue: nextValue) {
             field.wrappedValue = newValue
-        }
-    }
-
-    /// Hides a view based on the specified value.
-    ///
-    /// NOTE: This should only be used when the view needs to remain in the view hierarchy while hidden,
-    /// which is often useful for sizing purposes (e.g. hide or swap a view without resizing the parent).
-    /// Otherwise, `if condition { view }` is preferred.
-    ///
-    /// - Parameter hidden: `true` if the view should be hidden.
-    /// - Returns The original view if `hidden` is false, or the view with the hidden modifier applied.
-    ///
-    @ViewBuilder
-    func hidden(_ hidden: Bool) -> some View {
-        if hidden {
-            self.hidden()
-        } else {
-            self
-        }
-    }
-
-    /// Conditionally applies the given transform if the given condition evaluates to `true`.
-    /// - Parameters:
-    ///   - condition: The condition to evaluate.
-    ///   - transform: The transform to apply to the source `View`.
-    /// - Returns: Either the original `View` or the modified `View` if the condition is `true`.
-    @ViewBuilder
-    func `if`<Content: View>(_ condition: @autoclosure () -> Bool, transform: (Self) -> Content) -> some View {
-        if condition() {
-            transform(self)
-        } else {
-            self
         }
     }
